Add a heterogeneous pools mode
Some trap configurations make one type of trap work so well by itself
that evolution has a hard time to come up with a better layout requiring
a specific configuration of other traps.  However if an already decent
layout is provided as a starting point the algorithm can improve upon it.
The idea here is to use pools with different configurations to provide
such seed layouts for the main pool to work off of.