Big accuracy rework for layout updates
The approximation based on fixed HP intervals was found wanting with more
complex layouts, to the point of uselessness at the default accuracy.
Increasing accuracy to useful levels slowed the program to a crawl.

Simulate() now calculates the maximum HP value which won't change any of
the result variables.  Threat and income calculations use it to perform
exact integration.  It is somewhat slower than the previous default
accuracy, but much faster than the accuracy which was required for useful
results.