Wrap cout stream operators with console
On Ubuntu on Windows Subsystem for Linux (at least, maybe elsewhere), when
a line (or the screen) is cleared, and then replaced with shorter output, the
old output would remain visible after the new string terminated.  e.g. when a
core mod is replaced with one with a shorter name while using --fancy.

    "Core: common/runestones:10"    when changed to
    "Core: common/fire:25"          would result in
    "Core: common/fire:25nes:10"

This would fix the problem by tracking the number of written characters, and
writing spaces to the end of the line when endl is encountered.