Rearrange the spire database to store cores in a separate table
Since delta values are no longer stored, there's no layout-specific data
in the mods.

This also fixes the deletion of obsolete layouts with cores.  The SQL was
looking for any core mods with lower values, without checking if they
applied to that layout.

Required C++ version is now c++17 because of std::optional.